% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost-data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult humans}
\format{A data frame with 32560 rows and 15 variables:
\describe{
  \item{Adm.clerical}{unknown}
  \item{Bachelors}{person is a bachlor}
  \item{Male}{gender}
  \item{Never.married}{did person marry?}
  \item{Not.in.family}{is person a part of a family}
  \item{State.gov}{state}
  \item{United.States}{is from the united states}
  \item{White}{is white}
  \item{X..50K}{unknown}
  \item{X0}{unknown}
  \item{X13}{unknown}
  \item{X2174}{unknown}
  \item{X39}{unknown}
  \item{X40}{unknown}
  \item{X77516}{unknown}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Adult/}
}
\usage{
adult
}
\description{
A dataset containing adult population personal details
}
\keyword{datasets}
