% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{methods}
\name{show,Deepboost-method}
\alias{show,Deepboost-method}
\title{Print method for Deepboost model
Evaluates a trained deepboost model object.}
\usage{
\S4method{show}{Deepboost}(object)
}
\arguments{
\item{object}{Object of class "Deepboost"}
}
\description{
Print method for Deepboost model
Evaluates a trained deepboost model object.
}
\details{
Prints :
Model error: X"
Average tree size: Y"
Number of trees: Z"
}
\examples{
dpb <- deepboost(y ~ .,
 data.frame(x1=rep(c(0,0,1,1),2),x2=rep(c(0,1,0,1),2),y=factor(rep(c(0,0,0,1),2))),
 num_iter=2,tree_depth=2)
print(dpb)
}
