% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.predict}
\alias{deepboost.predict}
\title{Predicts instances responses based on a deepboost model}
\usage{
deepboost.predict(object, newdata, type = "terms")
}
\arguments{
\item{object}{A Deepboost S4 class object}

\item{newdata}{A data.frame to predict responses for}

\item{type}{Type of prediction : "terms" - for class labels, "response" for probabilities}
}
\value{
A vector of respones
}
\description{
Predicts instances responses based on a deepboost model
}
\examples{
dpb <- deepboost(y ~ .,
 data.frame(x1=rep(c(0,0,1,1),2),x2=rep(c(0,1,0,1),2),y=factor(rep(c(0,0,0,1),2))),
 num_iter=2,tree_depth=2)
deepboost.predict(dpb,data.frame(x1=rep(c(1,1,1,0),5),x2=rep(c(1,1,1,1),5)))
}
